Ray Tracing Pseudocode. Ray, Color, Image
Ray Tracing Pseudocode. Ray, Color, Image

+++ obl2/obl2/colors.h
@@ -3,22 +3,22 @@
 #define COLORS_H
 
 struct Color {
-	float red;
-	float blue;
-	float green;
-	float alpha;
+	double red;
+	double blue;
+	double green;
+	double alpha;
 };
 
 // COLORS
 // -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-const Color BLACK = { 0.f, 0.f, 0.f, 1.f }; // RGBA
-const Color WHITE = { 1.f, 1.f, 1.f, 1.f }; // RGBA
-const Color RED = { 1.f, 0.f, 0.f, 1.f }; // RGBA
-const Color BLUE = { 0.f, 1.f, 0.f, 1.f }; // RGBA
-const Color GREEN = { 0.f, 0.f, 1.f, 1.f }; // RGBA
-const Color CYAN = { 0.f, 1.f, 1.f, 1.f }; // RGBA
-const Color MAGENTA = { 1.f, 0.f, 1.f, 1.f }; // RGBA
-const Color YELLOW = { 1.f, 1.f, 0.f, 1.f }; // RGBA
-const Color LIGHT_GREY = { 0.3f, 0.3f, 0.3f, 1.f }; // RGBA
+const Color BLACK = { 0., 0., 0., 1. }; // RGBA
+const Color WHITE = { 1., 1., 1., 1. }; // RGBA
+const Color RED = { 1., 0., 0., 1. }; // RGBA
+const Color BLUE = { 0., 1., 0., 1. }; // RGBA
+const Color GREEN = { 0., 0., 1., 1. }; // RGBA
+const Color CYAN = { 0., 1., 1., 1. }; // RGBA
+const Color MAGENTA = { 1., 0., 1., 1. }; // RGBA
+const Color YELLOW = { 1., 1., 0., 1. }; // RGBA
+const Color LIGHT_GREY = { 0.3, 0.3, 0.3, 1. }; // RGBA
 
 #endif
